Fix circular import of ImageList by moving it to util
Otherwise we get a circular import such as:
 File "/home/travis/build/metabrainz/picard/test/test_mbxml.py", line 6, in <module>
    from picard.metadata import Metadata
  File "/home/travis/build/metabrainz/picard/picard/metadata.py", line 24, in <module>
    from picard.coverart.imagelist import ImageList
  File "/home/travis/build/metabrainz/picard/picard/coverart/__init__.py", line 28, in <module>
    from picard.coverart.providers import cover_art_providers, CoverArtProvider
  File "/home/travis/build/metabrainz/picard/picard/coverart/providers/__init__.py", line 168, in <module>
    from picard.coverart.providers.local import CoverArtProviderLocal
  File "/home/travis/build/metabrainz/picard/picard/coverart/providers/local.py", line 26, in <module>
    from picard.coverart.image import CoverArtImageFromFile
  File "/home/travis/build/metabrainz/picard/picard/coverart/image.py", line 33, in <module>
    from picard.script import ScriptParser
  File "/home/travis/build/metabrainz/picard/picard/script.py", line 27, in <module>
    from picard.metadata import Metadata
ImportError: cannot import name Metadata
